edited money collecting

+++ Assets/Scripts/moneychunk.cs
@@ -4,5 +4,7 @@ using UnityEngine;
 
 public class moneychunk : MonoBehaviour
 {
-    public float value;
+
+    public int value;
+
 }
